requirements.txt
python-decouple
GoogleBard
browser-cookie3
dbus-python